Support IncludeUUID during write path
Support was removed due to TTL not being handled by it. It now re-added.
Actually the IncludeUUID on write path also use the cache like write
without UUID and when cache miss is also query Cassandra.
Doing so allow to both ensure the UUID was generated by SquirrelDB and
fetch the current TTL in Cassandra.
We still have a performance improvement on using the UUID instead of all
labels because we will send a write with one label (the UUID) instead of
multiple labels.