Add new "batcher" that rely on a WAL instead of Redis
This is still a WIP, as cluster is not yet implemented. There is also
few value that should be configurable (flush time for example).
This is working for single node.

For cluster, the idea is using a memberlist and split metrics by shard.
Each shard will be assigned to one SquirrelDB instance. All read and
write will be sent to this instance.
Since the write WAL is stored in Cassandra (and quickly, currently after
10 seconds), losing the current master don't lose more than 10 seconds
of data.
The way the WAL is purged makes that if two SquirrelDB think being
master for the same shard don't result in data lose but will result in
additional works.