Fix high lock contention in index
Index used to took lookupIDMutex (assumed to be a very fast lock, used
on *every* write request) and then took newMetricLock, which is a slow
cluster lock.
In addition the slow lock, when failed wait a least 10 seconds before
retry. This means that all writes requests where blocked for 10 seconds.