Add a debug endpoint to verify & fix index
On my development DB, I had issues in the index. The issues are likely
due to the DB not being dropped between commit and an older commit may
used incorrect code.
So the fix feature should not be needed anymore. The verify will allow
to confirm that current code is correct and don't produce issues.

The verify should be runnable on production DB, it only consume time
(and send lots of query to Cassandra) but not memory. The fix will block
all new metrics creation during the whole process (which easily took
minutes)