Add support for limit of number of points/series evaluated
This limit only apply to PromQL (not the remote read endpoints) and
limit the number of points/series evaluated not the number returned.
For example "sum_over_time(node_cpu_seconds_total[300d])" will likely
evaluate a lots of series and a lots of points but only return one serie
and few points.