Fix various bug due to race-condition in Redis cluster
The code assumed that Redis operation were sequential and atomic but
this isn't the case when using cluster.
On cluster, the pipeline could not be atomic (crash of one Redis during
the pipeline execution) not sequential (the pipeline may cover multiple
redis).
This resulted in unexcpeted situation where metric could have an
flushDeadline but not data. Which caused SquirrelDB to try taking-over
some metrics but failing to do because excluded due to no data.

Now Redis will fill struct with empty points instead of not filling
anything.
Also use two sequential command during removal of metric. First remove
it (expire it) then remove it from known-metrics. This will allow to
recover based on known-metrics.