Detect and fix race-condition on metric deletion
If a metric is deleted (due to expiration, it's the only way to delete a
metric) another SquirrelDB having cache entry for that metrics could
re-use it.
Add two measure to avoid using invalid cache:
* don't use cached value when the metrics is close to expire or have
  expired. This will ensure that such cache entry are refreshed from
  Cassandra.
* When we try to update the expiration date, if we didn't find the
  metric ID in previous expiration date (means another SquirrelDB did an
  update), drop the cache.

There is still some small window for race-condition (which require
re-use of ID, re-use of same previous expiration date, ...). To avoid
them we should probably need to take the newMetrics lock which we want
to avoid.