Store aggregated timestamp scaled
Since Gorialla TSZ encoding used milliseconds on uint32, it didn't
allowed to store more than 49 days in one partition.
Since we want the default parition size to be 80 days, it was an issue.

By scaling the value with the resolution (5 minutes, or 300 000
milliseconds), we can store 40000 years which is enough :)