Use alpine for base image of SquirrelDB
Using alpine for Docker image allow to have the sleep command.
Due to the way Kubernetes to upgrade (concurrently stop the POD and remove the endpoint
from service), to have graceful upgrade we need to ensure POD stopping isn't too
fast. A sleep is the easiest hack to achieve this.