Enable tombestone purge on TSDB data tables
When using TimeWindowCompactionStrategy, by default Cassandra never re-compact
sstable older than the compaction_window. This means that even if 90% of the rows
could be dropped due to TTL, the file isn't compacted to drop them.
Since SquirrelDB allow to write data with different TTL, we need to re-compact older
sstable to clear expired rows.

This is enabled with "tombstone_threshold".
We also need to set "unchecked_tombstone_compaction" because SquirrelDB write use
one partition for 5 days while the Cassandra compaction_window is 6 days. Due to this
at least one partition always spawn across two sstable files. unchecked_tombstone_compaction
allow to process the file anyway.
But because of unchecked_tombstone_compaction, a sstable which crossed the tombstone_threshold
might have nothing droppable (Cassandra don't drop if the partition is also present in newer
sstable). That means the output file of the compaction might also cross the threshold.
To avoid a compaction loop, we set tombstone_compaction_interval to not re-process the
file for the next 7 days, which is enough for more data to be dropable (the partition in
the newer sstable might be dropped at that time).