Fix few issue with index_verify
We used to don't return hadIssue=true when we found issue on non-shared part of
the index (we just printed them, but don't set the boolean).
Also we miss reported that some ID were not in global All IDs posting,
when it was that this ID had some errors.
This means that an ID with minor issue (e.g. not yet being purge due to
expiration) would be reported on errors in all shards.