Use AddN/RemoveN everywhere AND fix bug from the mutation they do
AddN and RemoveN mutate the list passed as argument. This resulted
in the bulkDeleter to possibly have the IDs to delete being mutate which
could result in some IDs not being processed.
The consequence is that some IDs would be left in the following place:
* index_id2labels
* the global all IDs posting
* the per-shard present and maybe present IDs posting

It would correctly be removed from other postings and labels2ID, which means that
* The ID can't be reused (still present in global all IDs). No conflict possible
* The ID can't be returned by search (not present in postings)
* If a metric with same labels is re-wrote, a new metric ID will be allocated (normal behavior).

So the only issue is that an ID is lost and won't ever be reused.