Avoid wrongly deleting labels2id entry
During a partial write (Cassandra issue when a metric is created), we could
end with a metric ID being allocated and present in id2labels but
not in labels2id.
This situation is okay as the first write will fail (error returned to
HTTP caller) and fixed on next write retry. But the next write will use
another ID.

Later on, when the first ID expire we shouldn't delete the labels2id
that point to the second ID.

We used to do this which means that after expiration of the first ID,
the label2id entry is dropped which would result in a new metric being
created (if the metric is still wrote).
Consequences should be limited to duplicating the metric IDs because
postings (used in Search - that is when reading) aren't touched.