Fixed wrong existance list being used
Shard tasks a list of metric IDs present in themself. We have two lists:
* maybe precence
* and precense

The former one is only filled at end of metric addition to a shard (and
removed first). This one is used for list of metrics that are actually
present.
The first one (maybe presence) is used to track a possible partial write
(Cassandra failure while adding the ID to the shard). During expiration
we use the maybe presence list to decide whether a shard need to be
processed.

We used to wrongly use the non-maybe list during expiration, which
resulted in no re-retry expiration in the shards if Cassandra had an
issue during the expiration.