Add a second test for locks
The existing tests mostly test that best performance is nearly achived by using gorouting
(simulate an action that is concurent inside one SquirrelDB, like metric creation).
This is done due to an optimization that use a Golang lock to pass the Cassandra lock between
gorouting.

The added test for to have a single gorouting per process (simulate an action that is a singleton
inside one SquirrelDB, like metric expiration). This should stress more the Cassandra lock part.