Fix SquirrelDB shutdown issue
Because the connection manager gorouting used the context provided by
caller to New(), it was possible to have that gorouting that completed
the shutdown() before Close() was called.
This could result in call to Session() to try to acquire a new session
while the manager is shutdown, which isn't expected.
This had two consequences:
* An (innocuous) error "Trying to close a non-existing session" could be
   shown
* The Cassandra Session could be closed which some component (Index,
   TSDB...) might still expect to have a working Cassandra session.