Use a separate lock for ConnectionObserver
Without a separate lock, a dead-lock might occur between
openSession (which hold the lock) and the observer that could
be called in the same gorouting that call gocql.Session().

This usually happen when Cassandra use a Cluster and that cluster
is starting up. ./run-test.sh cluster should produce this issue
often.