Allow lookupLabels to fail on some labels
This could happen when the metric is deleted at the same time it's read.
While this is unlikely in real usage (it means you are reading an old metrics
which no longer had points), the previously behavior was to fail the whole Search().
Since the Search() could return both old metric and still alive metrics, we shouldn't
file the whole request.
Now we just drop the metrics that is deleted.