Apply shard expiration at the same time as metric expiration
Previously when we updated metrics expiration, we also processed
the update of the shard expiration. But the way is was done implied
than the associated metric for shard expiration would be moved to
a expiration day-list on next execution.
This wasn't optimal, because it's very likely that we already needed
to update that day-list due to normal metric, so doing the shard
exporation metric at the same time is better.
In addition it make test a bit more complex as they would need to
call RunOnce an additional time.