Add cache to TSDB point reading
The cache works on TSDB reads (after index). It works by keeping the
last result in memory (which is anyway loaded in memory, it was just
kept a little shorter time), and reuse this last result if possible.
Since only the very last result is kept, any request that result in
more than one metric won't be cached.