Add test trying to produce the labels mutation bug
We needed to add a Copy() of labels from lookupLabels because in some case
labels seems to be mutated (label __name__ being set to "").
We were able to reproduce it on a running SquirrelDB but not on another one,
nor in tests, even if this test try to mimic the (likely) cause.