Avoid dropping points that are in the future.
Because of https://github.com/prometheus/prometheus/pull/14304 , points that are 10 minutes or more in the future are blocked by Prometheus. To prevent that inconvenient behavior, we subtract an offset from these timestamps so they're considered as below the 10-minute barrier. Then, once this checked is done, we re-add the subtracted offset to the timestamps and continue the normal processing.