fix: wrong point timestamp when data in the future exists
To support writting data with timestamp in the future, SquirrelDB could
temporary mangle write requests (add 48 hours before processing through Prometheus
and remove those 48 hours before writting to TSDB).
But when both data in the future and data in the present/past were present in the
same write requests, SquirrelDB wrongly removed 48 hours to data in the present/past
without doing the add 48 hours for them.
Now when a write request contains any data in the futures, the 48 hours is added to
all data.